Get updateIndices script to actually, er, update indices
At least for the article case; blog posts and snippets still to follow.

Note that we're trying to avoid updating parts of the index that don't
need updates by remembering the last-updated index. This may not even
be necessary because even a full index update is fast (about a couple
seconds).