Hook up real Redis-backed connection
This is a bit of a hack because `connectionFromPromisedArray` wants a
fully materialized array, and we explicitly don't want to do that, as it
defeats the purpose of pagination.

I'll replace this with something else later on, but for now hacking
around it with `new Array()` to produce an appropriately padded array.
Tested with:

```
query MyQuery {
  # Demo of viewer field.
  viewer {
    id
    name
    articles(first:5, after: "YXJyYXljb25uZWN0aW9uOjk=") {
      edges {
        cursor
        node {
          id
          title
          createdAt
          updatedAt
          tags
        }
      }
    }
  }

  # Proof that node() field works.
  node(id: "QXJ0aWNsZTpDb3B5aW5nIGxvY2FsIGZpbGUgaGllcmFyY2hpZXMgdXNpbmcgcnN5bmM=") {
    id,
    ...ArticleFragment
  }
}

fragment ArticleFragment on Article {
  id
  title
  createdAt
  updatedAt
  tags
  body {
    raw
    format
    html
  }
}
```