Fix exit status of updateSchema.js
Because this is async and nobody is awaiting it at the top level, this
was just hanging forever. We need to `exit` on success.

At the same time, I cleaned up the error pathway, actually aborting
early. I went with `process.exit(1)` rather than throwing an error
because we don't need a stack trace here; it would only clutter up the
output.