Use an enum type for markup format
This doesn't really buy us a hell of a lot right now, other than
expressing our intent, which I suppose is good enough reason to do it.

Specifically, if the back end returns an invalid format look "foo", the
`format` field in `markup{format}` will get nulled out. But note that we
don't do validation in the `resolve` function for `markup{html}`, other
than a `throw` for unrecognized types; this is what we were already
doing, and we're not actually leveraging the type information here, so
we're not getting any additional benefit at this point.