Make markup{format} non-nullable
Now if the back-end returns a invalid format, instead of just returning
null we also emit an error. Alas, it's not super informative, but I
guess it is better than nothing:

> Cannot return null for non-nullable field Markup.format.

I would have preferred something that mentions "format must be one
of...", but oh well.

Includes schema update for some other formats that I forgot to include
previously.