Remove pointless --date-order
Commits are _always_ "topologically ordered" in Git in the sense that a
parent will never be listed before its children, irrespective of
committer or author dates. The command line switches for customizing
order only change the behavior in "tie-break" conditions (ie. when a
commit has multiple parents).

So, in that case:

- `--date-order`: breaks ties by committer date.
- `--author-date-order`: breaks ties by author date.
- `--topo-order`: follows each line of history separately, such that
  commits from two parallel branches are never interleaved.

Of course, I should have known this, because I did know that you can't
do a global sort by any kind of date without traversing the entire
history, which could be very expensive.

None of this is at all relevant to us, as we only have a single, linear
branch to traverse. Get rid of the option.