Turn snippet titles into links
Not sure if the schema is the right place for this to reside, seeing as
it means we have routing information dispersed across the schema and the
front-end.

But if we don't expose this, we end up exposing the internal ID anyway,
because the client will need it if it is the one constructing the URL.

So, on the balance, I suspect this is probably ok.