Add static mirror of legacy content
Unlike the stuff on the "content" branch, this stuff will never be
updated, so it belongs on this branch with the other static stuff under
"public/".

This mostly works. Note that some if the pages have CDN URLs for the
stylesheets, so those would work even if I didn't copy the CSS over, but
I copied them all over anyway.

As noted, still need to get "foo" -> "foo.html" static handling working;
might need to make a custom middleware for that, or just delegate that
responsibility to nginx.

As I've moved the `express.static` middleware out of the conditional, I
can drop my custom handling of the "/" path.

Still to fix:

- Image URLs on product pages.
- YARD docs are busted (eg. for wikitext gem).