First crude cut at tags indexing
This just gives us a ranked list by taggings count. For some reason,
reading in the wiki articles is slow (snippets, on the other hand, are
fast. No idea why.).

Wanted to get this in while it's in a working state anywaFirst crude cut
at tags indexing

This just gives us a ranked list by taggings count. For some reason,
reading in the wiki articles is slow (snippets, on the other hand, are
fast. No idea why.).

Wanted to get this in while it's in a working state anywayy