Add "tag cloud"
Not actually a cloud as that feels pretty early-2000s to me right now.
Should change the name.

In fact, should probably review all of it. This was coded up mostly via
copy pasta while sitting in the walk-in closet of a Santa Cruz beach
house where we're staying for the weekend, while waiting for my daughter
(who was using said closet as a room) to go back to sleep.