Strip out __DEV__ code in prod
Note that this only applies to the bundle.js file. `gulp build` will
turn ES201* files in `src/` into ES5 files in `dist/`, but those don't
get affected by the plug-in config. `dist/bundle.js`, however, does, in
addition to going through Babel.

So:

  if (__DEV__)

becomes:

  if (process.env.NODE_ENV !== 'production')

via Babel. And that becomes:

  if ("production" !== 'production')

via the `DefinePlugin`. Then the `UglifyJsPlugin` considers that to be
equivalent to:

  if (false)

and strips it out.