Provide configuration for content repo
This gives us more flexibility for setting things up in prod. With this
set-up, we can pass in a custom location for the content repo, relative
to the root of the code repo.

For example, if we pass "." then the code repo _is_ the content repo.

In prod, however, we might want it to be an entirely separate thing. Say
the running code repo is at `/var/masochist/deploy/current`, then we
could pass "../../content" to have the app use the repo at
`/var/masochist/content` instead.