Remove references to MySQL
I've decided to go full "stateless" so that I can spin up a new EC2
instance, configure it with Ansible, and go. This is possible because I
make the Git repo the source of truth for all data, including tags (yes,
this really should be a static site, but anyway).

Note: this is a bad shrinkwrap because I forgot to use the `--dev` flag.
I'll fix that in a follow-up commit, because I am about to remove
nodegit as well.