Make wiki article lookups partially case-insensitive
For simplicity, the old site was entirely case-insensitive for wiki
articles (but not for other lookups like blog permalinks); from memory,
the way I did this was by setting per-table case-sensitivity options in
the MySQL database.

The new site was case-sensitive everywhere (because Git is
case-sensitive). Note that this was true even on OS X, where the
filesystem is not case-sensitive.

With this commit we switch to the Wikipedia style of case-handling: the
first letter is considered to be case-independent but the rest is not.
So, given an article like "Open source", we can link to it as "open
source" or "Open source" according to its position within a sentence:

> [[Open source]] is great! I really like [[open source]].

But note that "OPEN SOURCE" is not valid.

I originally wanted to do this by transforming the filename into a
pattern (ie. "[Oo]pen source"), but `git-ls-tree` doesn't support that,
even though other Git commands (like `git-log`) do (it's been that way
for a long time; see:
http://comments.gmane.org/gmane.comp.version-control.git/197735).

I also looked at doing this "cleverly" to reduce repetition, but it
wasn't much shorter and was harder to read, so I wrote it out the dumb
longhand way. For reference, the alternative was something like:

    possibleNames = [
      String.prototype.toUpperCase,
      String.prototype.toLowerCase,
    ].reduce((names, fn) => names.push(...getFilenamesWithExtensions(
        subdirectory,
        fn.apply(file[0]) + file.slice(1)
    )), []);

Fixes: https://github.com/wincent/masochist/issues/33