Normalize case in wiki titles
The previous commit made both "[[Foo]]" and "[[foo]]" work as links to a
wiki article, "Foo", but we were showing either "Foo" or "foo" in the
page title instead of normalizing it. This commit fixes that, making
sure that we display "Foo".

Along the way, fixed a minor bug to make the calls to
`getTimestampsCacheKey` consistent between the `updateIndices.js` script
and the `loadContent` module. With this, both will create the key using
"Foo" instead of "Foo" and "content/wiki/Foo" respectively.

Related: https://github.com/wincent/masochist/issues/33