Pull nodeDefinitions out into separate file
Stealing the "registry" idea from:

https://gist.github.com/taion/d161a58b9f7381d8fa9c

This enables us to break what would otherwise be a circular dependency
requiring all types implementing the node interface to be in the same
file as the node definitions themselves (because the definitions depend
on the types, and the types include the interface).

It does introduce the evil of a side effect at module evaluation time
due to the `registerType` calls, but it's probably worth it in order to
be able to split things up, especially as the schema grows in size.

In subsequent commits I'll keep chipping away at the schema breaking it
into smaller pieces.