Add code for rendering a 500 page
As noted inline in the comment, we can't render `<App />` yet because
we'll need some webpack magic in place so that we can handle CSS loaders
and such. Not going to do that yet because I am not even sure if this is
a good idea (might just want to return blank-ish 500 status and have
nginx serve up a static 500 page).