Style-ize and consolidate error pages
Now the 404 and 500 pages both go through the same code path (both
rendering to static markup, with no client-side piece), and beware: here
be dragons.

In order to get `<App>` and friends rendering on the server, we
conditionalize our CSS imports so that they only happen in the browser
(with the jank `process` check).

Might revert this if I get scared of the dragons.