Make links work in server-rendered static pages
I should look into this more, but `<Link>` renders as a non-link on the
server, so this wrapper swaps it out for `<a>` instead.

Might make all `<Link>` instances go through this for consistency in a
future commit, but haven't done it yet as I am not sure this is the
right approach.