Make SVG backgrounds work on Firefox
Sigh. As I noted when I originally implemented these colored SVG
backgrounds using masks, cross-browser support for this is still crappy,
years after WebKit rolled this out.

The most promising alternative technique that I tried was switching to
inline elements, and using a webpack "raw" loader to get the string of
the SVG file and dump it into a React component that dangerously set the
inner HTML. From there it is trivial to target the SVG with CSS, as it's
in the DOM at that point.

Alas, this sucked for two reasons:

- I can't use inline elements for external link icons, as that HTML is
  generated elsewhere.
- This relied on webpack, and I've already seen how that makes my
  server-rendered static pages freak out.

So, going with the dumbest possible but also most robust workaround:
hard-coding the fill color in the SVGs themselves. Thus is fulfilled the
prophecy from commit bbbe0bb4e9. This is the worst possible solution to
this issue, except for all other solutions.

Fixes https://github.com/wincent/masochist/issues/16