Add missing getKey call in Tag model
We were manually constructing a key name here rather than using the
abstraction provided by `getKey`. This is bad, and would have broken in
an upcoming commit where I add a cache-invalidation mechanism by
including a version number in the keys.

So, fix the call-site for now; in a later commit I am going to make this
whole `getKey` thing fool-proof by baking it into the client wrapper
transparently.