Add cache invalidation mechanism to Redis code
In f8c164d we fixed a bug in the Redis indexing code. We need a way to
invalidate the bad index data, and this commit provides it. Note that
the approach taken here is compatible with our deployment strategy: we
can do an effectively "zero down-time" switch to a new version, or a
rollback to an old version, because the two indices are entirely
independent.

Note that the current index in prod is only 295K, which is to say,
nothing, so we don't worry about expiry or cleaning up old indices. I
could deploy once a day, invalidating the cache each time, and still
only use up 100M in a year, so this is fine; if it later becomes a
problem I can add a cron job to prune out old cache versions.