Make sure canonical attribute has correct case for wiki articles
Do a round trip via the GraphQL schema to ensure we use the right
URL capitalization in the `rel="canonical"` attribute. This is only
needed for wiki articles, because only they have variable case. Alas, it
does slow us down a bit, making the rendering of index template
asynchronous, but it's only on initial page load (not on transitions)
and I think correctness is important here (otherwise there is no point
in even having a canonical attribute).

This is simpler for Wikipedia because they always redirect to the
uppercase form, even if the article title calls for lowercase.

Fixes: https://github.com/wincent/masochist/issues/51