Add Facebook comment plugin
Now, I'm not sure I actually will go through with this, but I do want to
give it a try. I've also included a `<Like>` component in this commit
but I am not rendering it anywhere because:

- I don't like the UI clutter
- I don't have server rendering or OG tags yet, so any FB stories that get
  generated will have shitty previews

As such, I've got substantial duplication between that `<Like>` and
`<Comments>` components which I haven't bothered to extract out into a
common `<FBSocialPlugin>` component.

Other reasons why I am not sure whether I will go with this:

- Concern about perf impact of including the SDK
- Concern about the visual appearance of the page with the plugin on it
- Concern about potential race conditions in FB SDK intialization and
  React, especially in the context of the single-page nature of the
  application

So I am going to commit this for now, think about it some more, and
decide whether to revert or go ahead with it.

Related: https://github.com/wincent/masochist/issues/44