Teach wikiserve about Redis cache versions
This is the companion change to 30985a5, which did the corresponding
thing for JavaScript.

Keeping these in sync is going to be manual, I think, because even
though they are in the same repo and I could make them pull from a
shared configuration file, I can't really make it truly shared because
the app and the microservice don't get deployed in lockstep but rather
independently.