Get rid of buggy `split()` calls
Splitting on colons won't work well when the item on the right side of
the split itself contains a colon.

This was leading, for example, for the page at /tags/rebase to not show
the wiki article at /wiki/git_rebase:_you're_doing_it_wrong, because it
would try to load the non-existent content at /wiki/git_rebase.

So, replace the `split()` call. This will fix the population of the
timestamps cache in `bin/updateIndices` (relatively harmless, as we'd
end up populating it correctly later on anyway). Bump the cache breaker
to force re-population. Finally, the change in `TagType` is the one that
actually fixes the aforementioned failure to load.