Circumvent race on search page
When we added `<Progress>` as a `renderLoading` component on the search
route we introduced a bug:

- Searching would call `this.props.relay.setVariables`, which would kick
  off a load.
- `<Search>` would get swapped out for `<Loading>`, unmounting it.
- Relay data would arrive, triggering callback on `<Search>`.
- `<Search>` would get `setState` on unmounted instance, causing React
  to emit a warning.

The quick fix applied here is to not render `<Progress>` on this route,
meaning it doesn't get unmounted and there is no race. We do this
because otherwise we won't see the "Search" -> "Searching..." button
change in the UI, which is intended to fulfil the "loading indicator"
purpose of the `<Progress>` component; in effect, `<Progress>` is
redundant here.

We also apply the deeper fix of adding an `ifMounted` wrapper that
shields us from doing the wrong thing. This, for example, will also
protect us from the case where we initiate a search and then trigger a
route transition away by clicking on another route. Note that we can't
use the `isMounted` React API because it is deprecated and doesn't
support ES6 class components anyway.

This is verbose, so I'd like to come up with a nice way of abstracting
this pattern out before rolling it out to the other places in the UI
that we do a `setState` in an async Relay callback.