Serve many static assets via the CDN
This doesn't buy us much as we are so light on assets right now anyway,
but I already have Cloudfront set up so I may as well use it.

We don't have enough assets to make asset sharding super useful, but
we're doing it based on filenames for:

- static bundled resources (JS and CSS)
- images going through the Markdown parser

Notably, not doing it for wikitext markup because that's deprecated and
hard to do without changes to the wikiserve service (because
`img_prefix` is not something we can override at parse time) and also
because we couldn't do per-image processing to choose shards; rather all
images would be handled by the same shard because we have no callback
available to us to inspect individual filenames and use them as inputs
into our hash function.

Closes: https://github.com/wincent/masochist/issues/60