Add `ifMounted` to make Relay `setVariables` safe
Thought about a few patterns, but this is the best I can come up with.

Despite the belief that `isMounted` is an antipattern (see:
http://facebook.github.io/react/blog/2015/12/16/ismounted-antipattern.html),
I think it was probably a mistake to deprecate it. There _are_
legitimate use cases for it, and removing it from ES6 class components
forces those legitimate use cases to use a bunch of boilerplate like in
this commit, in order to disallow the potential smell of the bad actors.

Note that Relay async callbacks as modified in this commit pay a higher
price because there is no singular resource that corresponds to them
which I could acquire in `componentDidMount` and then free in
`componentWillUnmount` (like I'm doing in the `<Time>` component, for
instance). Because there may be multiple in-flight, I'd have to maintain a
queue of them, which would probably be even more boilerplate than I am
adding in this commit.

Closes: https://github.com/wincent/masochist/issues/61