Fix unwanted page reloads on some transitions
Some of my new markup has this structure:

  <div class="prerendered">
    ...
    <a href="/wiki/Git">
      <strong>Git</strong>
    </a>
    ...
  </div>

Hadn't noticed this before because all my old markup came from the
wikitext module, which doesn't nest tags inside "a" elements.

So was failing to be caught by our click handler (`event.currentTarget`
was the "strong" tag and not the "a" tag). Change the handler to walk up
looking for "a" tags as well on the way to the "prerendered" div.

Closes: https://github.com/wincent/masochist/issues/62