Add `getLoaders` module
Having call sites obtain individual loaders is fragile, because
traversing a GraphQL query may indirectly involve another loader (for
example, the `feed` action superficially only needs the `Post` loader,
but it may indirectly need the `Wikitext` loader and perhaps others
too).

As these are not hot code paths, there is no harm in instantiating the
full set of loaders every time we need any one of them, so this commit
introduces a unified place to call into and get a complete set.