Cache feed
Key is a combo of content hash plus an internal version number that we
can bump if we ever want to change the format or fix a bug in the
generation.

Note that I switched the format to always indent because I didn't want
to risk having varying content in the cache depending on whether I'm in
`__DEV__` or not.

Fixes: https://github.com/wincent/masochist/issues/67