Remove import.rb script
As I am unlikely to run it again. Even if I do a mass transformation on the repo
in the future (for example, porting all existing wikitext files to markdown), I
probably won't re-run the script; am more likely to cook up a new script and use
`git filter-branch` in some way. Likewise if I end up doing a mass-import of old
content that predates the Rails app.