Add support for persisted queries
This is a total hack but is the cheapest way I could think of to
implement this feature. This is a toy app, so we don't bother with
persisting to an actual database. Queries don't have unique IDs for each
version of the query; you request "PostsIndexQuery" and you get the
result for the current version, which should be fine as the queries
basically never change in this app and if they do, you get a validation
error or an unexpected `null` somewhere and you refresh the page.