Add wikitext to Markdown migration scripts
These are egregious hacks but were a lot of fun to cook up. I'll
probably write a blog post on the subject.

In brief:

- Prerequisites:
  - `pandoc`, installed via Homebrew.
  - `wikitext` gem, installed with `sudo gem install wikitext -n
    /usr/local/bin` (necessary workaround for macOS S.I.P. breaking gem
    installs...).
  - `unpack-content` NPM package installed with `-g` switch.

Ran it using:

    git filter-branch --index-filter \
      'PREV=$(map $(git rev-parse -q --verify $GIT_COMMIT^)); ~/bin/wiki-to-markdown-filter $PREV' \
      HEAD

Pre-rewrite history is now on "archives/content-2017-05-13" branch.

Will require a full re-index (bump "redisCacheVersion" in
"shared/common.json" and run "update-indices"; no need to redeploy
wikiserve because we shouldn't be hitting it any more).