script: commit final tweaks to import-filter script
So I got the import working, but needed a few tries. In one of my
experiments I rewrote the top 3 commits to be empty (whoops), and then
when I fixed the bug I forgot to fully reset before I ran
`git-filter-branch`. It takes many minutes to run, so I didn't want to
do it all over again; so, I had to cherry-pick those commits on top and
delete the messed up ones. This is way the few preceding commits have
author dates from last night but commit dates from just now.

Observations:

- By reading the `git-filter-branch` source code, I established that I'd
  need an (albeit empty) `--index-filter` in order do get it to make the
  commit that I'd set up in my `--parent-filter`.
- The other big change is the need to do 3-way index merges pretty much
  the whole way through the history because we have one of the following
  scenarios:

  (1) Initial commits: replayed as-is.

      A --- B --- C --- D --- E --->

  (2) Inserting commits for legacy content: eg. after C, insert X & Y.

      A --- B --- C --------- D --- E --->
                   \           \     \
                    X --- Y --- D'--- E'--->

  (3) "Rebasing" pre-existing commits on top of rewritten history:

      > --- E --- F --- G --- H --->
             \     \     \     \
        > --- E'--- F'--- G'--- H'--->

Without these three-way merges, we lose our previously applied changes.
Note that as soon as you rewrite one commit, you're doomed to doing
merges for every single subsequent commit. You can see the merges
happening in the calls to `git read-tree` with three tree-ish params,
which represent the base, the "original" and the "new" (for example,
these could be G, G' and H in the diagram above).

There is likely still crap to clean up in some of the old content, but I
am fine with making edits and bumping the modification dates for those
items.