Improve Flow coverage
Flow was freaking out at the ConcreteBatch/ConcreteFragment types, which
are not currently exported from the relay-runtime package. This meant
that it couldn't "see" *other* types exported in the __generated__
files.

This exposed some Flow issues, which I was mostly able to fix by making
things non-nullable. I am mostly confident of some of those (eg. the ones
that resolve to a literal object), but there are others which I am not
so sure about. Consider connections, which use nullable fields in some
places. I don't expect the data fetches to ever fail and if they do,
we're screwed and can't render anything anyway. Should I litter my code
with null checks for something that I don't ever expect to happen, or
leave them out and risk an extremely unlikely runtime error?

Closes: https://github.com/wincent/masochist/issues/114