Rationalize some "export default" statements
Want to avoid separate `export default X;` statements, except in places
where doing that will lead us to eliminate all references to `X` from
the file. If the file is called `X.js`, we want `X` to appear inline
inside it at least once, so that I can see an `import` of `X` elsewhere,
grep for "X", and find the file that defines it among the results.