Fix error thrown on first render of <Time> components
This bug has existed since the component was first written (see
8ddc8237b6395e9). My intention here was obviously to use the
module-local variable, but I didn't, incrementing the variable from
state instead, but `previousState` is not defined.

In testing this, I noticed that we often get `ttl` of `Infinity` for
older content, and `setTimeout` actually fires immediately for that, so
this was probably broken in other ways too.

Rejigged it to use real state (values to be rendered) instead of the
dummy `stateCounter`.