Unbreak some redirects with streaming rendering
Because streaming rendering doesn't actually start producing any output
until somebody tries to `read()` from the stream, our `catch` statement
in `main.js` on the server wasn't going to catch anything thrown during
a render, because the render doesn't actually happen until later.

So, move the redirection logic into the router action. This means we no
longer have access to the router itself, so we have to change the way we
do soft (internal) redirects: we have to communicate by throwing an
exception, which means we need to split `RedirectError` into two --
`ExternalRedirectError` and `InternalRedirectError` -- so that we can
distinguish between the two cases.

(Note to self: probably have to check to see whether RenderTextError
also needs the same treatment; IIRC it is already happening in the
router, so we're good.)