Don't depend on toGlobalId from graphql-relay
Ok to depend on this on the server, but not so great on the client,
where it pulls in graphql and a bunch of other heavy-weight crap. Doing
this drops the bundle size from about 1,591,591 bytes to about 524,450
bytes (pre-gzip).

This done with the help of the excellent source-map-explorer package:

    yarn global add source-map-explorer
    source-map-explorer bundle.js bundle.js.map