Reset yarn
Immediately after running `yarn upgrade-interactive` I blew away the
`yarn.lock` and ran a clean install to see how different the produced
result would be. Turns out quite a bit, I suspect mostly because of the
way it chose to pull down the new version of autoprexifer (which I had
skipped because it was a major version bump).