Speed up content loading and therefore indexing
Get rid of very expensive `rev-list` calls by maintaining an in-memory
index. This makes indexing way faster, and paginating too. Content
loading is expensive when the cache is cold, so you notice it most with
pagination when things are getting loaded 10 at time, and more so with
indexing where you are loading literally thousands of things during a
full re-index.