Get rid of extra retain clowniness
Fixed the root (route?) cause of the problem which was we were rendering
first with:

  <App>
    {children /* old view */}
  </App>

then:

  <App>
    <Progress />
    {children /* old view */}
  </App>

and finally:

  <App>
    {children /* new view */}
  </App>

React doesn't take too kindly to this. The old view sees:

- `componentWillUnmount`
- `componentDidMount`
- `componentWillUnmount`

Instead of rendering the `<Progress>` component by prepending it as a
new child in a top-level `render` call, we instead just re-render the
same shape with different props. With this in effect, we avoid the
redundant unmount/remount and the old view just sees:

- `componentWillUnmount`

No need for the awkward and kludly retain any more.