Circumvent garbage collection
Actually, decided I do want this after all because I'd like to make the
back button work without hitting the network again.

Note that we do have a request/response cache which means that in the
normal case we *don't* hit the network. However, in the pagination case,
if I paginate through 100 wiki articles, navigate away, and then back to
the wiki index, we will "hit" the network with the original query again
(fetching the first 10); this will be a cache hit, so there won't be an
actual network request, but it would be better if we just rendered the
100 articles we were previously showing. Will require some hackery to
get that to work but a prerequisite is that we have to have held onto
the data in the first place, so I'm adding back the retain.