Hack in quick-and-dirty variable preservation for back button
This is messy as hell but gets the job done for the ArticlesRoute. As
you can see, we have to cheekily get the fragment variables by snooping
because it's not easy to get them out of Relay (ok only because we only
ever show this component in one place, alone). And there's some nasty
threading through of state through the router. And the crowning turd
atop it all is the way we decide to avoid the network and do a
`readQuery` instead of a `fetchQuery` on the basis of these variable
overrides.

Will follow-up with doing this for the other pagination containers.