Add parallelForEach to updateIndices script
I want to debug an issue with the index script, but it is made a little
harder by the fact that we index four content types in parallel. Make a
`parallelForEach` helper that maintains the existing behavior, but can
be forced to run serially by changing:

    parallelForEach(collection, callback);

to:

    parallelForEach(collection, callback, 1);

By default, the final param defaults to the number of CPU cores
(typically 8), which means that all four content types will be indexed
in parallel, just like before. If the collection were bigger, the use of
`Promise.race` means that we'd run 8 jobs at a time, and whenever one
finished, we'd add another so that 8 would be running at all times.