Allow "type:blog" query alone to return results
I put the conditional in here only to preserve the old behavior, which
returned nothing for a query containing only "type:blog" or similar.

But the use of `every` here means if we have no other atoms to search
for, we'll return `true` without firing the callback and we won't
actually filter the corpus.

One minor wrinkle to bear in mind now that I've implemented
filename-based search: a piece of content will be considered a match if
all search terms appear in the title OR all search terms appear in the
body; it will not be considered a match if a subset of terms appear in
the title AND a subset appear in the body, even if together those
subsets in combination would constitute a full match. I think I'm going
to ignore that scenario as just too edge-casey to even worry about.