Work around unwanted history mangling
This solution is ugly so may want to clean it up later, but it is
necessary because the "history" package undoes my careful encoding,
breaking everything when using regex escapes.

We need to teach `<Link>`, then, to take two "to" props: one for display
purposes, and one for history purposes. Ugh.