Insist on exact match for tmux session check
`tmux has-session -t masochist` will return true in the presence of a
"masochist-pages" session (ie. it is a prefix match), which means we
will end up attaching to that instead of creating the desired
"masochist" session. We can avoid this by forcing an exact match instead
with `-t=masochist` instead.