Fix log format compatibility with recent Git
Our regex here was working on the server, where we are running Git
version 2.14.5, and we get `git log --raw` output that contains strings
like this:

    :000000 100644 000000000... 9773b7718... A      content/snippets/1157.md

On my local machine, however, running 2.19.0, the corresponding output
is:

    :000000 100644 000000000 9773b7718 A    content/snippets/1157.md

I was surprised to find no mention of this in the release notes, but
spelunking in the code at least reveals the `GIT_PRINT_SHA1_ELLIPSIS`
environment variable that can, for now at least, be used to resurrect
the ellipsis:

    https://public-inbox.org/git/20171203212743.10228-3-bedhanger@gmx.de/

In any case, the ellipsis is evidently on the way out, so let's just
tweak our regex to handle both variants. I'd rather do that than update
Git on the server and then find the regex breaks *again* in the future
because of a possible revert.