Adjust route resolution to better match Universal Router default
The default resolver expects a context object and params, and passes
them both on to the action function.

We were ignoring the params (jamming them into an unused `extra`
parameter), and stuffing them into the context object instead.

Let's not do that. This commit changes the parameter passing to match
the default.