Stash reference to Relay API on router
Seeing as the router is passed around everywhere and available in
context, expose the Relay API on there so that it can be accessed from
anywhere (instead of being buried in the client/app.js and
server/main.js top-level scopes only).

I may tweak the ordering and propagation of the parameters still, but
this is a reasonable intermediate step.