Work around @babel/polyfill accessing non-existent `global`
Seeing as I am using useBuiltins = 'usage' in my @babel/preset-env
config, in theory I could just remove the polyfill from the entry
section, but I am a bit wary of doing that due to its experimental
status. Also, the docs (https://babeljs.io/docs/en/babel-preset-env)
don't make it clear whether items like 'regenerator-runtime/runtime' get
handled adequately or not, so I might just wait this one out.

It seems certainly true that much of what is in the polyfill could be
skipped as I am pretty sure I am not using all of it:

  require("core-js/es6");
  require("core-js/fn/array/includes");
  require("core-js/fn/string/pad-start");
  require("core-js/fn/string/pad-end");
  require("core-js/fn/symbol/async-iterator");
  require("core-js/fn/object/get-own-property-descriptors");
  require("core-js/fn/object/values");
  require("core-js/fn/object/entries");
  require("core-js/fn/promise/finally");
  require("core-js/web");
  require("regenerator-runtime/runtime");

All of this prompted because of this check in the polyfill:

  if (global._babelPolyfill && typeof console !== "undefined" && console.warn) {
    console.warn("@babel/polyfill is loaded more than once...");
  }

  global._babelPolyfill = true;

Which is causing this in prod (only, for some reason):

  Uncaught ReferenceError: global is not defined